created NumberToWordHandle class and migrated methods from Numbertoword class
