+++ lib/index.d.ts
@@ -1,6 +1,7 @@
 import { english, kananda, telagu, marathi, hindi } from "./config/language";
 import { LAN, NumberToWordStyle, DecimalStyle } from "./config/interfaces";
-declare class NumberToWord {
+import NumberToWordHandle from "./classes/numberToWordHandle";
+declare class NumberToWord extends NumberToWordHandle {
     private _lan;
     constructor(lan?: LAN);
     /**
@@ -11,27 +12,6 @@ declare class NumberToWord {
      * @returns string
      */
     getWord(_number: number, _style?: NumberToWordStyle, _decimalStyle?: DecimalStyle): string;
-    /**
-     * This converts the decimal number into words
-     * @param _number decimal number
-     * @param _decimalStyle "Currency" | "Scientific"
-     * @returns decimal in string
-     */
-    private DecimalNumberWord;
-    /**
-     * This retuen whole number  in words
-     * @param _number
-     * @param _style "LakhsAndCrore" | "MillionAndBillion"
-     * @returns string
-     */
-    private WholeNumberWord;
-    /**
-     * This function converts number into comaseparted string array
-     * @param _number  "LakhsAndCrore" | "MillionAndBillion"
-     * @param style
-     * @returns string[]
-     */
-    private convertComaSepartedArray;
     /**
      * This function converts number into comaseparted string
      * @param _number  "LakhsAndCrore" | "MillionAndBillion"
