+++ test/.yalc/multilingual-number-to-words/yalc.sig
@@ -1,1 +1,1 @@
-bd93e040fbb928083febf12760a4e3fe
+30fc0acfcedab3ba055434230a75fe51
