+++ test/yalc.lock
@@ -2,7 +2,7 @@
   "version": "v1",
   "packages": {
     "multilingual-number-to-words": {
-      "signature": "bd93e040fbb928083febf12760a4e3fe",
+      "signature": "30fc0acfcedab3ba055434230a75fe51",
       "file": true
     }
   }
